+++ app.config.js
@@ -73,6 +73,15 @@ export default {
 			isDefault: false
 		}
 	},
+	// 全局异常码，可以自定义提示结果
+	globalErrorCode: {
+	  // 阿里云10秒非正常超时，其实请求还在执行（且一般已经成功了，但前端接接受不到成功结果）
+	  "cloudfunction-unusual-timeout": "请求超时，但请求还在执行，请重新进入页面。",
+	  // 请求超时（真正的请求超时）
+	  "cloudfunction-timeout": "请求超时，请重试！",
+	  // 不在预期内的异常（如数据库异常、云函数编译异常等）
+	  "cloudfunction-system-error": "网络开小差了！"
+	},
 	// 页面风格
 	pageStyle: {
 		// 表单组件和表格组件的size