+++ common/function/myPubFunction.js
@@ -7,6 +7,7 @@ var myfn = {};
  * vk.myfn.test1();
  */
 myfn.test1 = function(obj = {}) {
+	let vk = uni.vk;
 	console.log("执行了自定义公共函数test1");
 	return obj;
 };